Fix functional tests for USE_NEUTRON
If USE_NEUTRON is True, we should be using the NeutronFixture
and not starting the nova-network service or stubbing out
fake networks in the nova database.

This makes that change and fixes the fallout in busted tests.

Change-Id: I9b35ed7497db8bd1eb74f4bb89631aabbcfeec0d
(cherry picked from commit c5ea3fb41bc36f0f07e7bef55c2bb67770a03d36)