Add support for getting volume details with a specified microversion
This allows us to get volume details and pass through a microversion.
This will be used to get a volume with microversion 3.48 where
the shared_targets and service_uuid parameters are in the response.
We will then use that information to determine if we should lock
attach/detach operations on that volume.

Part of blueprint multi-attach-volume

Change-Id: Ief30adfef35eefe03e5a3d3d9623cd48e4d334a4