libvirt: don't log snapshot success unless it actually happens
Because of where this message was in the finally block, it
would log that we successfully extracted a snapshot even
if we failed, like if _live_snapshot fails because the guest
is deleted concurrently during the snapshot process.

This moves the info log message into the block of code where
we actually know that we successfully extracted the snapshot.

Change-Id: Ie12592c2daecbe764fa52dde6b0dafdbcafe102e