Only call numa_fit_instance_to_host if necessary
The numa_fit_instance_to_host method emits a debug message
when there is no provided host or instance topology:

  "Require both a host and instance NUMA topology to fit instance on host"

This can be confused with something that is wrong when most of the
time it just means the request spec doesn't have a NUMA topology
requirement.

This change avoids calling the method altogether if there is no
host and instance topology, just like in the NUMATopologyFilter.

Change-Id: Iba0f616927dc44929df07a19f8f33d4cc07017d1