Allow force-delete even if task_state is not None
As of now, if vm task_state is not 'None', and user tries
to force-delete instance, then he gets HTTP 500 Error
and instance deletion doesn't progress.

The same is not the case, when user tries with delete api
instead of force-delete api, even if vm task_state is not 'None'.

Fixed the issue by allowing force-delete to delete instance
in task_state other than None.

Change-Id: Ida1a9d8761cec9585f031ec25e5692b8bb55661e
Closes-Bug: #1741000