Fix fake libvirt XML generation for disks
The fake libvirt driver generates disk xml for disks of various types,
but assumes file type for the source attribute. This causes a parse
failure for block devices.

No tests currently rely on this. I encountered it whilst writing a test
which did use a fake block volume, and it was a severe headscratcher
when the resulting parsed configuration had an empty source_path: being
a block device it was looking for source/dev, not source/file. I
ultimately didn't use that test, but would like to save somebody the
hairloss in the future.

Change-Id: I5cf594cb0cc8310ccfa4f31d981803ce152a69d5