Simplify logic in get_enforcer
The get_enforcer method is used by oslopolicy CLI scripts to generate
policy files. The scripts will use configuration files to find policy
files so that overrides can be generated with default values
registered in code. The get_enforcer method used to parse the
arguments passed in from the system and remove the `namespace` and
`output-file` arguments because they wouldn't be recognized while
processing configuration values.

This commit simplifies the logic of get_enforcer. A related
discussion was held in review:

  https://review.openstack.org/#/c/530828/3

Proposing this to nova since I attempted to use similar logic to
fix a problem in keystone, but figured the simplified logic might be
useful here, too.

Change-Id: I7cd27fe8c39ddfc6ec20f4cfe4d62912d4cebaa4