Fix race condition in retrying migrations
During testing in the alternate host series, it was found that retries
when resizing an instance would always fail. This turned out to be true
even before alternate hosts for resize was introduced. Further
investigation showed that there was a race in call to retry the resize
and the revert of the original attempt.

This patch fixes that race condition, and updates the regression test to
reflect the correct results.

Blueprint: return-alternate-hosts

Closes-Bug #1741125

Change-Id: I12b4e1d9e50fefd81128df85d75bb34564f9da2c