Add regression test for resizing failing when using CachingScheduler
The cold migrate task in conductor assumes that the scheduler (or compute)
created allocations against the source node for the instance when it
attempts to swap those allocations to the migration record. However, if
using the CachingScheduler, the scheduler doesn't create allocations in
Placement, and if the computes are >=Pike, they won't either, so conductor
fails to find the allocations for the instance on the source node and fails.

This adds a functional regression test to show the failure. A follow up
patch with the fix will modify the test to show it passing again.

Change-Id: I80a401f6adce1c4e77a595928d6b9f085ff769a8
Related-Bug: #1741307