Handle no allocations during migrate
The CachingScheduler does not create allocations in the
scheduler, so the assertion in the conductor migrate task
that the instance will have allocations on the source node
is incorrect. This removes the assertion and changes the
error to a debug message.

The related functional regression test is updated to show
resize working with the CachingScheduler again.

There are other places in the code, especially the compute
and resource tracker, where we log errors if we can't find
allocations and those likely need to be updated at some
point, but it can happen in a follow up.

Change-Id: I0bb2933c4ed7ed479206c9b06be7e30a2ec92f2a
Closes-Bug: #1741307