Add regression test for resize failing during retries
During testing in the alternate host series, it was found that retries
when resizing an instance would always fail. This turned out to be true
even before alternate hosts for resize was introduced. Further
investigation showed that there was a race in call to retry the resize
and the revert of the original attempt.

This adds a functional regression test to show the failure. A follow up
patch with the fix will modify the test to show it passing again.

Related-Bug: 1741125

Change-Id: I0760f3e3bc447c39a6443afd7857046a08b8f01d