Pass multiattach flag to reserve_block_device_name
This is similar to how we use reserve_block_device_name
for tagged attach in that we can fail fast in the API if
trying to attach a multiattach volume to an instance on
a compute host that can't support it, either because the
compute service is too old or the virt driver doesn't
support the multiattach capability yet.

A later change will build on this to pass the multiattach
flag from the API when a multiattach volume is requested
in a new microversion, similar to how 2.49 works.

Part of blueprint multi-attach-volume

Change-Id: Ieec2bbd8a23c861c89bf9922517fa6a5562f1937