[libvirt] Filter hypervisor_type by virt_type
ComputeNode.supported_hv_specs field has hypervisor_type
infomation about which type (e.g. KVM or both KVM and QEMU) of guests
are supported by the hypervisor, but it doesn't supply information
about which virt_type is set to that compute node.

This means that we have no way to know which type of guests are going
to be built in that compute node from the ComputeNode object. This can
cause wrong behavior in scheduler filters using the supported_hv_specs
object, such as ImagePropertiesFilter.

In libvirt driver, this patch adds a filter to exclude supported
hypervisor_types other than the virt_type which has been set by
operator via nova.conf.

Partial-Bug: #1195361
Change-Id: I986fe09f97e4d5cad940bea85a87c58606314dba