Add access_url_base to console_auth_tokens table
Unfortunately this column was missed off when console_auth_tokens
was added to the database. This change adds it in.

The access_url is needed for a protocol check in the console proxies.
Console proxies run per cell and do not run on compute hosts and thus
do not have access to the access_url config options in the compute
hosts nova.conf files.

So, we need to store part of the access_url in the database. We cannot
store the entire access_url because it contains the unhashed token, but
we can keep the base part of the url and generate the full access_url
on demand. An access url base looks something like this:

  http://127.0.0.1:6080/vnc_auto.html

and using it, we can generate the full access_url on demand for the
protocol check in the console proxies.

partially-implements: blueprint convert-consoles-to-objects

Change-Id: I0f672f5667d42b67d869ff9f467dbb64eb6c9ff9