libvirt: don't attempt to live snapshot paused instances
When we changed the default value of the
workarounds.disable_libvirt_livesnapshot config option value
to False in 980d0fcd75c2b15ccb0af857a9848031919c6c7d earlier
in Queens, we were testing against the Pike UCA packages which
has libvirt 3.6.0 and qemu 2.10. Live snapshots of a paused
instance work with those package versions as shown by the
test_create_image_from_paused_server test in Tempest.

However, if you just use the Ubuntu 16.04 packages for libvirt
(1.3.1) and qemu (2.5), that test fails and the live snapshot hangs
on the paused instance.

This change adds PAUSED to a list of power states that aren't
valid for live snapshot. We can eventually remove this when we
require (or add a conditional check for) libvirt>=3.6.0 and
qemu>=2.10.

Change-Id: If6c4dd6890ad6e2d00b186c6a9aa85f507b354e0
Closes-Bug: #1741667