Add exception to no-upcall note of cells doc
The cells v2 layout documentation clearly states that there are no
upcalls from cells back to the central API services. This mislead
me for sometime as I could not fathom how a compute node in a cell
was supposed to report its resource info.

It turns out nova looks up the placement service in the keystone
catalogue and contacts it directly which to my mind is an upcall. I
wonder if the author of the not felt that the placement service is
not really part of nova?

Change-Id: If14be8b182f0af4e4e6641046fec638c07e26546
Closes-Bug: #1742421