Increase notification wait timeout in functional tests
In the baseline the wait_for_notification() functional test util
function by default waits maximum 1 second for a notification to
arrive. Based on recent gate failures it seems that it is not enough
in some cases for the notitification to arrive. Therefore this patch
bumps the default timeout to 10 seconds. This will only make the test
execution slower if the notifiication arrives later or does not even
arrive. So in the happy case the test execution will not be slower.

Change-Id: I75e2bcd87535de7e44bdf317c3ab12ba8952dcca
Closes-Bug: #1742688