Handle swapping to a multiattach volume
Since swap volume doesn't attach the new volume using the
DriverVolumeBlockDevice.attach flow like normal attach or
server create, the new volume connection_info doesn't get
the "multiattach" value set in it, which is later read in
the virt driver's swap_volume method to build the disk device
config.

In the case of libvirt it relies on that multiattach flag
being set in the connection_info dict to set the 'shareable'
element on the LibvirtConfigGuestDisk object.

This change modifies the swap volume flow to perform the same
checks as in DriverVolumeBlockDevice._volume_attach. In this
change, for simplicity, it's just a copy/paste hack, and a TODO
is left to move this into common code later.

Part of blueprint multi-attach-volume

Change-Id: I173c404a54d1306e95c1cc3fbd651b137cb1c6aa