XenAPI/Stops the migration of volume backed VHDS
This commit aims to correct problems with the resize_up codebase that allows
the snapshot and migration of volume backed VDI/VHDs.  Since these are empty
stub disks, and the XenAPI does not allow these VDIs to be snapped, this results
in an SR_OPERATION_NOT_ALLOWED or similar error on attempt.

This change adds a check into the _process_ephemeral_chain_recursive method to
run the current userdevice through volume_utils.is_booted_from_volume.  To
achieve this, the method has been opened in scope to accept custom user_device
objects.  In a future commit we will need to rename this method for clarity
and correct its dependancies that call it.  I have added a TODO for this to be
done by myself. The check will ensure that the userdevice is not volume backed
and then continue to snapshot and migrate the disk as needed, else increment
and move on.

Closes-Bug: #1745072
Change-Id: I7cd2977c8268c1f73062b5d0b2b68ea686db99fe