Make sure that functional test triggered on sample changes
To be able to define different irrelevant-files for the functional jobs
than the ones defined in openstack-zuul-jobs we need to copy the jobs to
the nova tree and modify the fields in tree.

Technically we could factor out the irrelevant-files regexp list from
functional and functional-py35 jobs as they are the same today. However
in the future when they diverge we cannot simply override the
irrelevant-files in one of the jobs. Therefore this patch does not
introduce a common base job for functinal and functional-py35 jobs
to discurage trying to override.

The openstack-tox-functional and fuctional-py35 are removed from the
nova part of the project-config in
I56d44f8dff41dbf3b2ff2382fa39b364f55f9a44

Conflicts:
      .zuul.yaml

NOTE(mriedem): The conflict is due ot the nova-lvm job definition
not existing in Pike, which is OK because it's a non-voting job
that runs in the experimental queue so we don't need it in Pike.

Depends-On: https://review.openstack.org/538619/

Closes-Bug: #1742962
Change-Id: Ia684786d1622da7af31aa4479fc883a7c65848ff
(cherry picked from commit d4377c2d537ea998ed96cc8d7071877eefab994f)