ProviderTree.data => ProviderData
An instance of ProviderTree is used to cache resource provider
information in the report client.  We will also use a ProviderTree
instance as the input and output of ComputeDriver.update_provider_tree.
We will then need to compare those two and flush any differences back to
the placement API.  Thus, we need a way to access the details of
individual resource providers in a ProviderTree.

Prior to this patch, provider details were hidden behind a private
_Provider class, and only accessible from ProviderTree via methods that
assumed you already knew the thing you were trying to compare against.

Enter ProviderData, which has all the same attributes as _Provider, but
whose modification is discouraged in various ways (naming,
documentation, and exceptions attempting to set attributes).  The new
ProviderTree.data method returns a ProviderData of the specified
_Provider.

Change-Id: I832bda86d9a6f9924d28dcdbd7091ed466c19196
blueprint: nested-resource-providers