SchedulerReportClient.set_traits_for_provider
Introduce SchedulerReportClient.set_traits_for_provider, which accepts a
provider UUID and list of traits to push back to the placement API.  On
success, the local ProviderTree cache is updated appropriately.  On
failure, an exception is raised.  If the failure was a conflict (409),
the exception is a subclass of PlacementAPIConflict (see below) so that
the caller may decide to refresh and retry.

This change also introduces the _ensure_traits helper, which makes sure
a given set of traits exists in the placement service.

Per design [1], we want consumers of SchedulerReportClient to be able to
detect placement API conflict errors (HTTP 409), as distinct from other
errors.  Such consumers (the resource tracker) may respond to conflicts
by redriving the operation in progress, whereas other exceptions will be
allowed to bubble up.  PlacementAPIConflict will be the (base) class of
exceptions raised in response to placement 409s for which we want to
allow this kind of behavior.

[1] http://eavesdrop.openstack.org/meetings/nova_scheduler/2018/nova_scheduler.2018-01-08-14.00.log.html#l-251

Change-Id: I9d9b4a5cb5f6c5d255c00e22923728892c37383f
blueprint: nested-resource-providers
blueprint: ironic-driver-traits