Track associated sharing RPs in report client
With this change, SchedulerReportClient keeps track of any sharing
providers associated by aggregate with the compute node RP in the same
way as traits and aggregates.  Specifically:

- We use ProviderTree methods to save the information for each provider.
- We refresh sharing RP information along with aggregate and trait
  information whenever we "ensure" a provider exists.

Note that we also retrieve and track the aggregates and traits
associated with each sharing provider.

The refreshing (including cache expiry) of sharing RPs is done in the
same method with that of aggregate and trait associations.  The theory
being that all of these associations should change out of band with
roughly the same frequency (i.e. very rarely).

Change-Id: I840ebc7940e65af30e66d902bf360668ccb5a30e
blueprint: nested-resource-providers