Raise on API errors getting aggregates/traits
Previously in the SchedulerReportClient, _get_provider_aggregates and
_get_provider_traits would log a warning and return None on placement
API errors.  The callers were ignoring these errors, resulting in the
internal ProviderTree representation potentially being out of sync with
reality.

With this change set, these two methods raise exceptions on API errors.
As currently implemented, these errors will bubble up into the resource
tracker and compute manager in the same way as other unhandled
exceptions.  That is, they land in the trap in
update_available_resource_for_node, where they are logged and otherwise
ignored.

Why is this an improvement?  Because at least the routine that
encountered the error doesn't continue on as if nothing went wrong.  (We
could have dealt with this by having the callers check for None
explicitly, but this approach was deemed more rigorous and robust.)

Change-Id: I20fe0d895f61ec83759cf880a87238ecc43c08f5
blueprint: nested-resource-providers