SchedulerReportClient.set_aggregates_for_provider
Introduce SchedulerReportClient.set_aggregates_for_provider, which
accepts a provider UUID and list of aggregates to push back to the
placement API.  On success, the local ProviderTree cache is updated
appropriately.  On failure, an exception is raised.

Note that generation conflicts are not handled in this iteration of this
method, because the placement API does not yet support generations on
aggregate associations.  That work is forthcoming.

Change-Id: I4a7d4c1f3caaebebdb38ae04c5f85e57ec967852
blueprint: nested-resource-providers