placement: using the dict format for the allocation in claim_resources
In the placement API version 1.12, the dict format is consistent format in
the APIs `PUT /allocations/{consumer_uuid}`, `GET /allocations/{consumer_uuid}`
and `GET /allocation_candidates`. This patch switches to this consistent
format in the claim_resources method.

There is one upgrade scene which needs to be care. Since we send alternative
hosts to the compute node as Selection object. The Selection object includes
the json dict of allocation object. That means if a try of instance building
happened in the middle of upgrade, the old format allocation json dict will
send back to the new version cell conductor. So the claim_resources method
still needs to support old version allocation json format.

Change-Id: I4314ade528c9cb69bbf6c55d31566e1e188fef09