Set server status to ERROR if rebuild failed
Currently there is no indication that the rebuild was refused,
and worse, we may have a wrong imageref for the instance.

This patch set the instance to ERROR status if rebuild failed in the
scheduling stage. The user can rebuild the instance with valid image
to get it out of ERROR state and reset with right instance metadata and
properties.

Closes-Bug: 1744325

 Conflicts:
	nova/tests/unit/conductor/test_conductor.py

NOTE(melwitt): The conflict was because the select_destinations method
has additional parameters in Queens that don't exist in Pike.

Change-Id: Ibb7bee15a3d4ee6f0ef53ba12e8b41f65a1fe999
(cherry picked from commit d03a890a34f632adc9a19a33c8a5aebbccec50e4)