Report Client: PUT empty (not None) JSON data
Previously if a False-ish payload was sent to SchedulerReportClient.put,
we wouldn't send it through in the API call at all.  But False-ish
payloads may be legitimate: e.g. there is currently no DELETE API for
/resource_providers/{u}/aggregates so this is how you would remove all
aggregate associations for a provider.

In any case, placement's PUT API refuses to accept a request that
doesn't have Content-Type: application/json, which is set automatically
by Session if json is not None.

With this change set, we send payloads through to PUT unless they're
actually None.

Change-Id: I69d2b16d515590907ca0e0dc4da77dcf8e539976
Closes-Bug: #1744786