Don't filter out sibling sets with one core
libvirt collects and exposes information about the host just as it does
for guests. This information includes things like the host CPU topology.
Nova uses this information to, among other things, generate an instance
of 'nova.objects.numa.NUMACell' for each host NUMA cell. The 'siblings'
field for each of these objects is populated using a combination of the
siblings information exposed by libvirt minus the cores we've been told
not to use, via the nova-specific 'vcpu_pin_set' option.

On pretty much any standard (x86) machine, it would be exceedingly
unlikely to find an asymmetric topology like so:

  socket 0
    core 0
      thread 0  (CPU #0)
      thread 1  (CPU #2)
    core 1
      thread 0  (CPU #1)

(note the missing 'CPU 3').

However these do happen in nova's world and are perfectly valid, thanks
to the aforementioned 'vcpu_pin_set' option. Historically, once we've
generated this list of sibling sets needed for the 'siblings' field,
we've filtered out sets that contain less than two siblings (i.e. the
cores have no siblings, or CPU #1 in the above example). This would be a
fair assumption on normal hardware, where asymmetric topologies like the
above wouldn't exist, but it doesn't really make sense in this use case.
What we're trying to expose via the siblings attribute is the
relationships between cores: if no relationship exists, we should say as
much.

The code that populates the 'siblings' attribute is modified to no
longer exclude sibling sets with only one sibling. This allows us to
simplify code that was previously having to switch between the
'siblings' attribute and the flat 'cpuset' attribute. In addition, all
tests that use the 'NUMACell' object are updated to set the 'sibling'
attribute as it is now always set in "real" code. This ensures behavior
hasn't regressed.

Change-Id: I8982ab25338969cd98621f79b7fbec8af43d12c5
Partial-Bug: #1744965