Ensure emulator threads are always calculated
In change 'I071182c00a0', the '_get_pinning' inner function of
'nova.virt.hardware._pack_instance_onto_cores' was extended to calculate
emulator thread pinning information along with guest vCPU pinning
information. Unfortunately, there is a code path used for calculating
vCPU pinning information in '_pack_instance_onto_cores' which does not
use this function and emulator thread pinning was therefore not being
calculated in some cases.

Replace this code path with a call to '_get_pinning', ensuring that all
permutations follow a similar flow and that the emulator thread pinning
information is calculated, as expected.

We also add a log which appears to have been missed in the
aforementioned change.

Change-Id: I021ce59048d6292055af49457ba642022f648c81
Closes-Bug: #1744965