Expand on when you might want to set --max-count for map_instances
The default behavior for the "nova-manage cell_v2 map_instances"
command is to map all instances in the cell in batches of 50.

This can be slow when there are several thousand instances in the
deployment and an operator may want to specify a higher --max-count
value and run the command until it completes.

This simply updates the command option description and man page to
point this out for consideration.

Change-Id: I59c2ed89fe02212977445f6825c6da8fedbb8ccf
Related-Bug: #1742649