New-style _set_inventory_for_provider
Given the UUID of a provider, set the inventory records for the provider
to the supplied dict of resources.

Compare and contrast with the existing set_inventory_for_provider, which
was created for the simpler get_inventory code path from the resource
tracker.

This one is specially formulated for use by update_from_provider_tree,
because from there, whereas we DO need to _ensure_resource_class(), we
ALSO want to short out if inv_data matches what's already cached, but we
DON'T want to:
- _ensure_resource_provider
- refresh_and_get_inventory
- DELETE if inventory is empty, else PUT
- retry
- invalidate the cache on failure

This version is more in the spirit of set_traits_for_provider.

Change-Id: I45f1df6ca5618f76337319ba225493625a5ee2d6
blueprint: update-provider-tree