Reduce complexity of _from_db_object
This patch reduces the complexity of the helper method _from_db_object
by creating a second helper method to handle processing the instance's
extra attributes. Although this reduces the complexity of
nova/object/instance.py, it is not possible to reduce the max-complexity
in tox.ini, since there are still other files with a complexity of 34.

Reducing the complexity of _from_db_object means that it is now possible
to add new extra attributes without increasing the overall complexity
of the nova code.

Change-Id: Ide987bb5e15f1fe2eb61ad0b218d1d5cc6bf49a2