Make sure that we have usable input for graphical console
Graphical console is optional thing. But when it is enabled then it is
good to have some input devices.

On x86(-64) it is handled by PS/2 keyboard/mouse. On ppc64 we have USB
keyboard/mouse. On aarch64 we have nothing.

So make sure that we have USB Host controller available and that usb
keyboard is present. Also USB tablet (default pointer_model device) will
have a port to plug to.

Closes-bug: 1745340

Change-Id: I69a934d188446a1aa95ab33975dbe1d6e058ebf9